Move fetch house detail api to util
