+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\houseDetails.js":"1","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\index.js":"2","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\house.js":"3","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\Error.js":"4","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\Login.js":"5","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\errors.js":"6"},{"size":254,"mtime":1612033802373,"results":"7","hashOfConfig":"8"},{"size":637,"mtime":1612033802373,"results":"9","hashOfConfig":"8"},{"size":230,"mtime":1612033802373,"results":"10","hashOfConfig":"8"},{"size":255,"mtime":1612195771458,"results":"11","hashOfConfig":"8"},{"size":2398,"mtime":1612193393397,"results":"12","hashOfConfig":"8"},{"size":230,"mtime":1612191530160,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"suuiyq",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\houseDetails.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\index.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\house.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\Error.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\Login.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\reducers\\errors.js",[]]
+[{"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\HouseCard.js":"1","C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\containers\\HouseDetails.js":"2"},{"size":1568,"mtime":1612778475645,"results":"3","hashOfConfig":"4"},{"size":3017,"mtime":1612779143439,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"suuiyq",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\components\\HouseCard.js",[],"C:\\Users\\acer\\Desktop\\Microverse project\\final rails react\\find-your-house-frontend\\src\\containers\\HouseDetails.js",[]]
